+++ aiohttp_swagger3/__init__.py
@@ -1,4 +1,5 @@
 __all__ = (
+    "swagger_doc",
     "RapiDocUiSettings",
     "ReDocUiSettings",
     "RequestValidationFailed",
@@ -12,7 +13,7 @@ __version__ = "0.5.6"
 __author__ = "Valetov Konstantin"
 
 from .exceptions import ValidatorError
-from .swagger_docs import SwaggerDocs
+from .swagger_docs import SwaggerDocs, swagger_doc
 from .swagger_file import SwaggerFile
 from .swagger_route import RequestValidationFailed
 from .ui_settings import RapiDocUiSettings, ReDocUiSettings, SwaggerUiSettings